Added security configuration partially.

+++ src/main/java/com/craftsy/webapp/database/entity/UserRole.java
@@ -16,7 +16,7 @@ public class UserRole {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-   private Long id;
+   private Integer id;
 
     //many to one with users table
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
@@ -25,7 +25,7 @@ public class UserRole {
     private User user;
 
     @Column(name = "user_id" , insertable = false, updatable = false)
-    private Long userId;
+    private Integer userId;
 
     @Column(name = "role_name")
     private String roleName;
